+++ src/holiday_card/__init__.py
@@ -1,4 +1,4 @@
 """Holiday Card Generator - Create printable greeting cards."""
 
-__version__ = "1.2.0"
+__version__ = "1.3.0"
 __author__ = "Holiday Card Team"